Adding product controller API

+++ ShoppingProducts.Domain/Dto/ProductDto.cs
@@ -6,6 +6,6 @@ namespace ShoppingProducts.Domain
     {
         public required string Name {get; set;}
         public required string Description {get; set;}
-        public required double Price {get; set;}
+        public required float Price {get; set;}
     }
 }
